+++ src/tic-tac-toe/Square.tsx
@@ -1,12 +1,5 @@
 import * as React from "react";
-import { OnClick } from "./Game";
-
-export type OX = "X" | "O" | null;
-
-interface ISquare {
-  value: OX;
-  onClick: OnClick;
-}
+import { ISquare } from "./index";
 
 export const Square = ({ value, onClick }: ISquare) => (
   <button className="square" onClick={onClick}>
